Use the proper ServerSpec shared `spec_helper`
No need to transfer a shared `spec_helper` over using TK’s data_path
config. TK does the right thing using the `spec_helper` located at
`test\integration\helpers\serverspec`.
