+++ .kitchen.yml
@@ -7,7 +7,6 @@ driver:
 provisioner:
   name: chef_zero
   require_chef_omnibus: 12.4.3
-  data_path: test/shared
 
 platforms:
   - name: centos-5.11
