+++ test/integration/helpers/serverspec/support/omnibus_build.rb
